+++ packages/ludiek/src/engine/LudiekEngine.ts
@@ -5,7 +5,7 @@ import { LudiekEngineSaveData } from '@ludiek/engine/peristence/LudiekSaveData';
 import { LudiekConsumer, LudiekInput } from '@ludiek/engine/input/LudiekConsumer';
 import { LudiekOutput, LudiekProducer } from '@ludiek/engine/output/LudiekProducer';
 import { LudiekTransaction } from '@ludiek/engine/transaction/LudiekTransaction';
-import { ControllerSchemas, LudiekController, LudiekRequest } from '@ludiek/engine/request/LudiekRequest';
+import { ControllerSchemas, LudiekController, LudiekRequest } from './request/LudiekController';
 import { ControllerNotFoundError } from '@ludiek/engine/request/RequestError';
 import {
   BonusContribution,
