+++ packages/ludiek/src/engine/LudiekEngineConfig.ts
@@ -1,7 +1,7 @@
 import { LudiekEvaluator } from '@ludiek/engine/condition/LudiekEvaluator';
 import { LudiekConsumer } from '@ludiek/engine/input/LudiekConsumer';
 import { LudiekProducer } from '@ludiek/engine/output/LudiekProducer';
-import { LudiekController } from '@ludiek/engine/request/LudiekRequest';
+import { LudiekController } from './request/LudiekController';
 import { LudiekPlugin } from '@ludiek/engine/LudiekPlugin';
 import { LudiekModifier } from '@ludiek/engine/modifier/LudiekModifier';
 import { LudiekFeature } from '@ludiek/engine/LudiekFeature';
