+++ packages/ludiek/src/engine/output/LudiekProducer.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { LudiekDependencies, LudiekEngineContribution } from '../LudiekEngineContribution';
+import { LudiekDependencies, LudiekEngineContribution } from '@ludiek/engine/LudiekEngineContribution';
 import { IsNonEmpty } from '@ludiek/util/types';
 
 /**
